requirements.txt
# 加密货币价格预测系统 - 依赖包

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# HTTP请求
requests>=2.31.0

# 日志和配置
python-dotenv>=1.0.0

# AI模型SDK
google-genai>=0.2.0
openai>=1.0.0

# 可选: 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0
